+++ CHANGELOG.md
@@ -71,6 +71,11 @@ The major version bump is due to upping the required PHP version from `8.5` to `
 * Added the following inspectors to the pixel library:
 	- `ColorAt`
 	- `TopColors`
+* Added the following methods to the `Color` class:
+	- `Color::fromHsl()`
+	- `Color::fromHwb()`
+	- `Color::invert()`
+	- `Color::with()`
 * Added dotenv loader.
 * Added `Signer::validateOrThrow()` method.
 * Added support for all new Redis commands added in Redis 8.2, 8.4, 8.6, 8.8 and 8.10 to the Redis client.